+++ apps/api/Persistence/Repository/IPageable.cs
@@ -1,4 +1,4 @@
-namespace Application.Services;
+namespace Persistence.Repository;
 
 public interface IPageable
 {
